feat: refactored bridge
- Cleaned up implementation of bridge
- Removed unneeded babel-polyfill and webpack globals
