+++ client/coral-embed/src/SnackBar.js
