+++ client/coral-embed/src/onIntersect.js
@@ -1,20 +1,30 @@
 export default function onIntersect(el, callback) {
+  // Ensure that the intersection observer is available.
   if (!IntersectionObserver) {
     // tslint:disable-next-line:no-console
-    console.warn('IntersectionObserver not available');
+    window.console.warn('IntersectionObserver not available, rendering now');
     callback();
     return;
   }
-  const options = {
-    rootMargin: '100px',
-    threshold: 1.0,
-  };
 
-  const observer = new IntersectionObserver(entries => {
-    if (entries[0].isIntersecting) {
-      observer.disconnect();
-      callback();
+  // Create the Intersection Observer that will wait till the embed is within
+  // view and will then call the callback.
+  const observer = new IntersectionObserver(
+    entries => {
+      if (entries[0].isIntersecting) {
+        // Stop receiving intersection events.
+        observer.disconnect();
+
+        // Fire the callback.
+        callback();
+      }
+    },
+    {
+      rootMargin: '100px',
+      threshold: 1.0,
     }
-  }, options);
+  );
+
+  // Start observing the element for visibility.
   observer.observe(el);
 }
