+++ webpack.config.js
@@ -310,19 +310,32 @@ const applyConfig = (entries, root = {}) =>
     config,
     {
       entry: entries.reduce(
-        (entry, { name, path: modulePath, disablePolyfill = false }) => {
-          const entries = [
-            path.join(
-              __dirname,
-              'client/coral-framework/helpers/webpackGlobals'
-            ),
-          ];
-          if (disablePolyfill) {
-            entries.push(modulePath);
-          } else {
-            entries.unshift('babel-polyfill');
-            entries.push(modulePath);
+        (
+          entry,
+          {
+            name,
+            path: modulePath,
+            disablePolyfill = false,
+            disableWebpackGlobals = false,
           }
+        ) => {
+          // Create all the entries to be added to the final build target.
+          const entries = [];
+
+          if (!disablePolyfill) {
+            entries.push('babel-polyfill');
+          }
+
+          if (!disableWebpackGlobals) {
+            entries.push(
+              path.join(
+                __dirname,
+                'client/coral-framework/helpers/webpackGlobals'
+              )
+            );
+          }
+
+          entries.push(modulePath);
 
           entry[name] = entries;
 
@@ -350,7 +363,8 @@ module.exports = [
       {
         name: 'embed',
         path: path.join(__dirname, 'client/coral-embed/src/index'),
-        disablePolyfill: process.env.TALK_DISABLE_EMBED_POLYFILL === 'TRUE',
+        disablePolyfill: true,
+        disableWebpackGlobals: true,
       },
     ],
     {
